Central store of Chatbot functions in a hooks file
- ok so i learned that Mobile and Chatbot copied functions from each other, which was a lot of duplicate code.
 - Since I'm preparing to help make a mobile edition to the Planner page, I thought that we should consider refactoring the Navbar on that to store a unified version of a mobile Navbar (if we plan to expand mobile view to our other pages)
   - the real kicker is determining if the time it took me to refactor and store all functions pertinent to Chatbot is worth it. Idk, the real kicker is seeing what Lighthouse says (localhost was terrible)
