+++ src/components/ChatBotNavbar.tsx
@@ -9,6 +9,7 @@ import {
   DropdownMenuTrigger,
 } from "./ui/dropdown-menu";
 import { chatEventEmitter } from "../utils/chatEventEmitter";
+import { useChatbot } from "../hooks/useChatbot";
 
 interface Message {
   role: "user" | "bot";
@@ -24,16 +25,26 @@ interface Conversation {
   conversation_name?: string;
 }
 
-// These are in milliseconds
-const CONVERSATIONS_CACHE_EXPIRATION_TIME = 1000 * 60 * 60;
-const CONVERSATION_CACHE_EXPIRATION_TIME = 1000 * 60 * 60;
-
 // check environment
 const ENVIRONMENT = import.meta.env.VITE_ENVIRONMENT as string | undefined;
 
 const ChatBotNavbar = () => {
   const { user, logout  } = useAuth();
   const location = useLocation();
+
+  const { 
+    conversations, 
+    conversation_id, 
+    error, 
+    loading,
+    setConversations,
+    deleteConversation,
+    renameConversation,
+    setConversationId,
+    initialLoad
+  } = useChatbot();
+
+
 
   const [isInWebapp, setIsInWebapp] = useState(false);
   const [profilePicture, setProfilePicture] = useState<string>("");
@@ -53,13 +64,9 @@ const ChatBotNavbar = () => {
   const [conversationToDelete, setConversationToDelete] = useState<string | null>(null);
   const [deleting, setDeleting] = useState(false);
 
-  // Mobile navbar state
-  const [conversation_id, setconversation_id] = useState<string | null>(null);
-  const [conversations, setConversations] = useState<Conversation[]>([]);
-  const [chatHistoryLoad, setChatHistoryLoad] = useState(false);
-  const [error, setError] = useState<string | null>(null);
-
-  const CRUD_API = import.meta.env.VITE_CRUD_API;
+  useEffect(() => {
+    initialLoad();
+  }, []);
 
   // check pfp 
   useEffect(() => {
@@ -109,146 +116,6 @@ const ChatBotNavbar = () => {
     return { todayChats, pastChats };
   };
 
-  // Delete conversation function - copied from ChatBot.tsx
-  const deleteConversation = async (conversationId: string) => {
-    if (!user?.uid) {
-      console.warn("User ID is missing. Cannot delete conversation.");
-      return;
-    }
-    setError(null);
-
-    try {
-      // Optimistically update cache
-      const cachedConversationsString = localStorage.getItem("chatbot_conversations");
-      if (cachedConversationsString) {
-        const cached = JSON.parse(cachedConversationsString);
-        if (cached?.data) {
-          const updatedCache = {
-            ...cached,
-            data: cached.data.filter((item: Conversation) => item.conversation_id !== conversationId),
-          };
-          localStorage.setItem("chatbot_conversations", JSON.stringify(updatedCache));
-        }
-      }
-
-      if (!CRUD_API) throw new Error("CRUD_API environment variable is missing.");
-
-      const token = await user.getIdToken();
-      if (!token) throw new Error("Failed to retrieve authentication token.");
-
-      const response = await fetch(CRUD_API, {
-        method: "POST",
-        headers: { "Content-Type": "application/json" },
-        body: JSON.stringify({
-          userId: user.uid,
-          action: "deleteConversation",
-          token,
-          conversationId,
-        }),
-      });
-
-      if (!response.ok) {
-        const errorText = await response.text();
-        throw new Error(`Failed to delete conversation: ${response.status} - ${errorText}`);
-      }
-
-      // Remove in state
-      setConversations((prev) => prev.filter((item) => item.conversation_id !== conversationId));
-
-      // Update localStorage cache too
-      const updatedCacheString = localStorage.getItem("chatbot_conversations");
-      if (updatedCacheString) {
-        const updatedCache = JSON.parse(updatedCacheString);
-        if (updatedCache?.data) {
-          const filtered = updatedCache.data.filter(
-            (item: Conversation) => item.conversation_id !== conversationId
-          );
-          saveConversationsToCache(filtered);
-        }
-      }
-
-      if (conversation_id === conversationId) {
-        setconversation_id(null);
-        localStorage.removeItem("chatbot_conversation");
-      }
-    } catch (err) {
-      const msg = err instanceof Error ? err.message : "Failed to delete conversation";
-      setError(msg);
-      console.error("Error deleting conversation:", err);
-    }
-  };
-
-  // Rename conversation function - matches desktop version exactly
-  const renameConversation = async (conversationId: string, newTitle: string) => {
-    if (!user?.uid) {
-      console.warn("User ID is missing. Cannot rename conversation.");
-      return;
-    }
-    setError(null);
-
-    try {
-      console.log('Mobile navbar: Starting rename for conversation:', conversationId, 'to:', newTitle);
-
-      // Optimistic state update - same as desktop
-      setConversations((prev) => {
-        const updated = prev.map((conv) => 
-          conv.conversation_id === conversationId 
-            ? { ...conv, title: newTitle, conversation_name: newTitle } 
-            : conv
-        );
-        console.log('Mobile navbar: Updated conversations:', updated.length);
-        return updated;
-      });
-
-      // Update local storage - same as desktop
-      const cachedConversationsString = localStorage.getItem("chatbot_conversations");
-      if (cachedConversationsString) {
-        const cached = JSON.parse(cachedConversationsString);
-        if (cached?.data) {
-          const updatedCache = {
-            ...cached,
-            data: cached.data.map((item: Conversation) =>
-              item.conversation_id === conversationId ? { ...item, title: newTitle, conversation_name: newTitle } : item
-            ),
-          };
-          localStorage.setItem("chatbot_conversations", JSON.stringify(updatedCache));
-          console.log('Mobile navbar: Updated localStorage cache');
-        }
-      }
-
-      if (!CRUD_API) throw new Error("CRUD_API environment variable is missing.");
-      const token = await user.getIdToken();
-      if (!token) throw new Error("Failed to retrieve authentication token.");
-
-      console.log('Mobile navbar: Making API call to rename conversation');
-      const response = await fetch(CRUD_API, {
-        method: "POST",
-        headers: { "Content-Type": "application/json" },
-        body: JSON.stringify({
-          userId: user.uid,
-          action: "renameConversation",
-          token,
-          conversationId,
-          newName: newTitle,
-        }),
-      });
-
-      if (!response.ok) {
-        const errorText = await response.text();
-        console.error('Mobile navbar API Error:', response.status, errorText);
-        throw new Error(`Failed to rename conversation: ${response.status} - ${errorText}`);
-      }
-
-      const result = await response.json();
-      console.log('Mobile navbar: Rename successful:', result);
-      
-    } catch (err) {
-      const msg = err instanceof Error ? err.message : "Failed to rename conversation";
-      setError(msg);
-      console.error("Mobile navbar: Error renaming conversation:", err);
-    }
-  };
-
   // Helper function to truncate text with ellipsis
   const truncateText = (text: string, maxLength: number = 24) => {
     if (text.length <= maxLength) return text;
@@ -268,7 +135,7 @@ const ChatBotNavbar = () => {
             className={`flex-1 text-left rounded-sm px-3 py-2 ${active ? "bg-secondary" : "bg-bglight"} transition-colors`}
             onClick={() => {
               // Update active conversation in mobile navbar
-              setconversation_id(conv.conversation_id);
+              setConversationId(conv.conversation_id);
               
               // Emit event to load conversation in main ChatBot component
               chatEventEmitter.emit('loadConversation', {
@@ -316,153 +183,6 @@ const ChatBotNavbar = () => {
       </li>
     );
   };
-
-  // ---------- helpers from chatbot.tsx ----------
-  const isCacheValid = (timestamp: number, cacheUserId: any, cacheValidFor: number): boolean => {
-    if (!user?.uid || !timestamp || !cacheUserId) return false;
-    const currentTime = Date.now();
-    return currentTime - timestamp < cacheValidFor && user.uid === cacheUserId;
-  };
-
-  const saveConversationsToCache = (convs: any[]) => {
-    localStorage.setItem(
-      "chatbot_conversations",
-      JSON.stringify({
-        data: convs,
-        timestamp: Date.now(),
-        userId: user?.uid,
-      })
-    );
-  };
-
-  const sortConversationsByDate = (convs: Conversation[]) => {
-    return convs.sort((a: Conversation, b: Conversation) => {
-      const aTime = new Date(a.messages?.[a.messages.length - 1]?.timestamp || 0).getTime();
-      const bTime = new Date(b.messages?.[b.messages.length - 1]?.timestamp || 0).getTime();
-      return bTime - aTime;
-    });
-  };
-
-
-  const fetchConversation = async () => {
-    if (!user?.uid) {
-      console.warn("User ID is missing. Cannot fetch conversations.");
-      return;
-    }
-
-    setChatHistoryLoad(true);
-    setError(null);
-
-    try {
-      const cachedConversationsString = localStorage.getItem("chatbot_conversations");
-
-      if (cachedConversationsString) {
-        const cachedConversations = JSON.parse(cachedConversationsString);
-        if (
-          cachedConversations.timestamp &&
-          cachedConversations.userId &&
-          isCacheValid(
-            cachedConversations.timestamp,
-            cachedConversations.userId,
-            CONVERSATIONS_CACHE_EXPIRATION_TIME
-          )
-        ) {
-          const cached = Array.isArray(cachedConversations.data) ? cachedConversations.data : [];
-          const processedConversations = cached.map((conv: Conversation) => ({
-            ...conv,
-            title: conv.title || conv.conversation_name || conv.messages?.[0]?.content || "Untitled Conversation",
-          }));
-          const sorted = sortConversationsByDate(processedConversations);
-          setConversations(sorted);
-          setChatHistoryLoad(false);
-          return cached;
-        }
-      }
-
-      if (!CRUD_API) throw new Error("CRUD_API environment variable is missing.");
-
-      const token = await user.getIdToken();
-      if (!token) throw new Error("Failed to retrieve authentication token.");
-
-      const response = await fetch(CRUD_API, {
-        method: "POST",
-        headers: { "Content-Type": "application/json" },
-        body: JSON.stringify({
-          userId: user?.uid,
-          action: "getConversations",
-          token,
-        }),
-      });
-
-      if (!response.ok) {
-        const errorText = await response.text();
-        throw new Error(`Failed to fetch conversations: ${response.status} - ${errorText}`);
-      }
-
-      const data = await response.json();
-
-      const convs: Conversation[] = Array.isArray(data)
-        ? data.map((conv: Conversation) => ({
-            ...conv,
-            title: conv.title || conv.conversation_name || conv.messages?.[0]?.content || "Untitled Conversation",
-          }))
-        : [];
-
-      const sorted = sortConversationsByDate(convs);
-      setConversations(sorted);
-      saveConversationsToCache(sorted);
-      return sorted;
-    } catch (err) {
-      const errorMessage = err instanceof Error ? err.message : "Failed to fetch conversations";
-      setError(errorMessage);
-      console.error("Error fetching conversation:", err);
-    } finally {
-      setChatHistoryLoad(false);
-    }
-  };
-
-
-  const initialLoad = async () => {
-    if (!user?.uid) return;
-
-    const cachedData = localStorage.getItem("chatbot_conversation");
-    if (cachedData) {
-      const { conversation_id, timestamp, cacheUserId } = JSON.parse(cachedData);
-
-      if (timestamp && cacheUserId && isCacheValid(timestamp, cacheUserId, CONVERSATION_CACHE_EXPIRATION_TIME)) {
-        setconversation_id(conversation_id || null);
-
-        const cachedConversationsString = localStorage.getItem("chatbot_conversations");
-        if (cachedConversationsString) {
-          const cachedConversations = JSON.parse(cachedConversationsString);
-          if (
-            cachedConversations.timestamp &&
-            cachedConversations.userId &&
-            isCacheValid(
-              cachedConversations.timestamp,
-              cachedConversations.userId,
-              CONVERSATIONS_CACHE_EXPIRATION_TIME
-            )
-          ) {
-            const cached = Array.isArray(cachedConversations.data) ? cachedConversations.data : [];
-            const processedConversations = cached.map((conv: Conversation) => ({
-              ...conv,
-              title: conv.title || conv.conversation_name || conv.messages?.[0]?.content || "Untitled Conversation",
-            }));
-            const sorted = sortConversationsByDate(processedConversations);
-            setConversations(sorted);
-            return;
-          }
-        }
-      } else {
-        localStorage.removeItem("chatbot_conversation");
-      }
-    }
-
-    // No valid cache — fetch list (won't auto-load a thread; new chat screen by default)
-    await fetchConversation();
-  };
-  // ---------- end helpers ----------
 
   // Route-based style
   useEffect(() => {
@@ -738,7 +458,7 @@ const ChatBotNavbar = () => {
               
               // Add to mobile navbar's conversation list
               setConversations(prev => [newConversation, ...prev]);
-              setconversation_id(newConversationId);
+              setConversationId(newConversationId);
               
               // Emit event to trigger new chat in main ChatBot component
               chatEventEmitter.emit('startNewChat');
@@ -750,7 +470,7 @@ const ChatBotNavbar = () => {
           </button>
 
           {/* Conversations */}
-          {chatHistoryLoad && <p className="text-textsecondary text-sm">Loading conversations...</p>}
+          {loading && <p className="text-textsecondary text-sm">Loading conversations...</p>}
           {error && <p className="text-destructive text-sm">{error}</p>}
 
           {Array.isArray(conversations) && conversations.length > 0 ? (
